Switching to another approach
Due to insufficient accuracy of geographical area calculations, we switched to another approach. This time we'll try to download rain-intensity-marked pictures from RainViewer API. They are of fixed size. Also, the most useful thing which this approach brings is the ability to calculate which are the exact corner coordinates of the area we are analyzing. Based on that, we can further calculate the total area show on the map which really is the purpose of switching to this approach.

The problems I encountered so far are deciding which tile of Google Maps we need to download (converting real world coordinates to Google Maps' XYZ) and calculating the actual area of the downloaded Google Maps tile.

maps_utils.py can solve the task of converting real world coordinates to Google's XYZ and obtaining its corner coordinates (top left, top right, bottom left, bottom right).
areacalc.py can calculate the area covered by a polygon using real world coordinates.

The biggest part of the work left is to get all the parts (Python files) to work together.